requirements.txt
Django==1.4.5
Pillow==2.0.0
django-appconf==0.6
django-forms-bootstrap==2.0.3.post1
django-jsonfield==0.8.12
eventlog==0.6.2
ipython==0.13.2
metron==1.1
pytz==2013b
six==1.3.0
wsgiref==0.1.2
coverage
django-coverage
pep8
pyflakes
# Django-guardian used for privacy settings
django-guardian

# Django-south for migrations
South
